have the consul client manage the lock session
When a consul lock is lost, there is a possibility that the associated
session is still active. If the client creates a new lock, the session
id won't match and the lock can't be recovered. Re-using the existing
lock can try and reuse the session, but the same external factors that
caused the loss of the lock may still cause the session to be lost.

First have the lock monitor retry the lock internally for at least 10
seconds (5 attempts with the default 2 second wait time). If the consul
lock can't recover itself, then cancel the session as soon as possible
(terminating the PreiodicRenew will call Session.Destroy), and start
over.